Use BigDecimal instead of Double for currency rate calculations (#134)
use BigDecimal instead of double for accuracy

+++ service/src/main/java/org/whispersystems/textsecuregcm/currency/FixerClient.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import org.whispersystems.textsecuregcm.util.SystemMapper;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
@@ -20,7 +21,7 @@ public class FixerClient {
     this.client = client;
   }
 
-  public Map<String, Double> getConversionsForBase(String base) throws FixerException {
+  public Map<String, BigDecimal> getConversionsForBase(String base) throws FixerException {
     try {
       URI uri = URI.create("https://data.fixer.io/api/latest?access_key=" + apiKey + "&base=" + base);
 
@@ -58,7 +59,7 @@ public class FixerClient {
     private String date;
 
     @JsonProperty
-    private Map<String, Double> rates;
+    private Map<String, BigDecimal> rates;
 
   }
